requirements.txt
# Core
numpy>=1.24
pandas>=2.0
scipy>=1.11

# Geospatial / ocean data
xarray>=2023.6
netCDF4>=1.6
cdsapi>=0.6          # ERA5 download via CDS API
motuclient>=1.8      # CMEMS download
cartopy>=0.22        # map projections and coastlines
shapely>=2.0

# ML
scikit-learn>=1.3
xgboost>=2.0
torch>=2.1
shap>=0.44

# Visualisation
matplotlib>=3.8
seaborn>=0.13
cmocean>=3.0         # perceptually uniform ocean colormaps

# Notebook
jupyter>=1.0
ipywidgets>=8.0

# Utils
tqdm>=4.65
python-dotenv>=1.0
